data/requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
pylint
geopandas
matplotlib
sqlalchemy
geoalchemy2
ipython
python-dotenv
psycopg2-binary
praw
nltk